Change harga type from Integer to Double

+++ src/main/java/ui/ft/ccit/faculty/transaksi/barang/Barang.java
@@ -17,7 +17,7 @@ public class Barang {
     private Integer stok;
 
     @Column(name = "harga")
-    private Integer harga;
+    private Double harga;
 
     @Column(name = "persen_laba")
     private Double persenLaba;
